requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
google-generativeai>=0.3.0
gtts>=2.5.0
torch>=2.0.0
pydub>=0.25.1
